+++ core/src/main/kotlin/no/elg/hex/island/Island.kt
@@ -4,7 +4,6 @@ import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.files.FileHandle
 import com.badlogic.gdx.utils.Queue
 import com.fasterxml.jackson.annotation.JsonAlias
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.fasterxml.jackson.annotation.JsonValue
 import com.fasterxml.jackson.module.kotlin.readValue
 import kotlinx.coroutines.CancellationException
@@ -66,7 +65,6 @@ import kotlin.math.max
 import kotlin.system.measureTimeMillis
 
 /** @author Elg */
-@JsonIgnoreProperties("initialLoad", "handRestore", "authorRoundsToBeat")
 class Island(
   width: Int,
   height: Int,